+++ common/src/main/java/net/mehvahdjukaar/supplementaries/SuppPlatformStuff.java
@@ -1,11 +1,12 @@
 package net.mehvahdjukaar.supplementaries;
 
-import com.simibubi.create.content.kinetics.mechanicalArm.ArmInstance;
 import dev.architectury.injectables.annotations.ExpectPlatform;
 import net.mehvahdjukaar.supplementaries.common.utils.SlotReference;
 import net.minecraft.core.BlockPos;
+import net.minecraft.core.Direction;
 import net.minecraft.core.particles.BlockParticleOption;
 import net.minecraft.server.level.ServerLevel;
+import net.minecraft.util.RandomSource;
 import net.minecraft.world.entity.Entity;
 import net.minecraft.world.entity.EntityType;
 import net.minecraft.world.entity.LivingEntity;
@@ -117,4 +118,13 @@ public class SuppPlatformStuff {
         throw new AssertionError();
     }
 
+    @ExpectPlatform
+    public static void tryBurningByFire(ServerLevel level, BlockPos pos, int chance, RandomSource random, int age, Direction direction) {
+        throw new AssertionError();
+    }
+
+    @ExpectPlatform
+    public static boolean canCatchFire(Level level, BlockPos pos, Direction direction) {
+        throw new AssertionError();
+    }
 }